Rename some symbols to read easily, fix close button's appearance for dark model.
