+++ ios/Sources/SettingFeature/SettingToggleItem.swift
@@ -1,0 +1,22 @@
+import SwiftUI
+import Styleguide
+
+struct SettingToggleItem: View {
+    
+    let title: String
+    @Binding var isOn: Bool
+    
+    var body: some View {
+        Toggle(isOn: $isOn, label: {
+            Text(title)
+        })
+        .toggleStyle(SwitchToggleStyle(tint: AssetColor.primary.color.color))
+        .listRowBackground(AssetColor.Background.contents.color.color)
+    }
+}
+
+struct SwiftUIView_Previews: PreviewProvider {
+    static var previews: some View {
+        SettingToggleItem(title: "ダークモード", isOn: .constant(true))
+    }
+}
